remove useless progn
The goal of the removed progn was to unconditionally execute the list of
statements and return a non-nil value. It can be safely removed,
because (skip-chars-backward) always return non-nil, so the (if ...)
form will always be executed. Moreover, (if ...) will also return
non-nil, unless (backward-word) fail because of buffer edge which we
would like to detect anyway.